adding apiserver tet
